% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_predictor_dataframe}
\alias{build_predictor_dataframe}
\title{Constructs a dataframe collecting the predictors from a formula}
\usage{
build_predictor_dataframe(form)
}
\arguments{
\item{form}{A formula.}
}
\value{
A dataframe.

Each column of the dataframe is the vector of values associated with a
predictor in `form`.
}
\description{
`build_predictor_dataframe` searches the environment in which a
formula was defined and collects the vectors associated with the
predictors of that formula into a dataframe.
}
\examples{
\dontrun{
library(RGOB)
A <- c(1,2,3,4)
X <- c(5,6,7,8)
Z <- c(1,0,1,0)
build_predictor_dataframe(A ~ X + Z)
# returns: 
# X Z
# 5 1
# 6 0
# 7 1
# 8 0
}
}
