% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_predictor_dataframe_from_data}
\alias{build_predictor_dataframe_from_data}
\title{Constructs a dataframe collecting the predictors from a formula}
\usage{
build_predictor_dataframe_from_data(form, data)
}
\arguments{
\item{form}{A formula.}

\item{data}{A dataframe.}
}
\value{
A dataframe.

Each column of the dataframe is the vector of values associated with a
predictor in `form`, as extracted from `data`.

If the formula has a dot as a predictor (e.g. Y ~ .), then the entire
dataframe `data` is returned minus the column containing the outcome.
See the examples.
}
\description{
`build_predictor_dataframe_from_data` extracts a dataframe containing
the predictors in a formula, from a larger dataframe.
}
\examples{
\dontrun{
library(RGOB)
dat <- data.frame(A = c(1,2,3,4)
                  X = c(5,6,7,8)
                  Z = c(1,0,1,0))
build_predictor_dataframe_from_data(A ~ X + Z, dat)
# returns: 
# X Z
# 5 1
# 6 0
# 7 1
# 8 0

build_predictor_dataframe_from_data(A ~ ., dat) # equivalent to previous call.
}
}
