% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{build_regression_data}
\alias{build_regression_data}
\title{Constructs a dataframe for use in regression functions}
\usage{
build_regression_data(form, Y, X)
}
\arguments{
\item{form}{A formula.}

\item{Y}{A vector.}

\item{X}{A vector or a dataframe.}
}
\value{
A dataframe.
}
\description{
`build_regression_data` construct a dataframe `dat` from
`Y` and `X` for use in a regression such as `lm(form, data=dat)`.
}
\details{
The names of the columns of the dataframe constructed must
correspond to the variables named in `form`, except in the case
where the formula `form` has a dot on the right-hand-side
(e.g. Y1 ~ .). If `X` is already a dataframe, not check is
done. If it is a vector, then it is converted into a dataframe.
}
\examples{
\dontrun{
A <- c(1, 2)
B <- c(3, 4)
C <- c(5, 6)

X <- data.frame(X1 = B, X2 = C)

build_regression_data(Y ~ X1, A, B)
# returns:
# Y X1
# 1 3
# 2 4 

build_regression_data(Y ~ ., A, B)
# returns:
# Y X
# 1 3
# 2 4

build_regression_data(Y ~ X1 + X2, A, X)
# returns:
# Y X1 X2 
# 1 3  5  
# 2 4  6

build_regression_data(Y ~ X1, A, X)
# returns:
# Y X1 X2 
# 1 3  5  
# 2 4  6
}

}
