% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{.build_pred_fn}
\alias{.build_pred_fn}
\title{Constructs a prediction function from a regression output}
\usage{
.build_pred_fn(reg_res, pred_names)
}
\arguments{
\item{reg_res}{The output object from a regression, e.g. `lm` or `glm`}

\item{pred_names}{The names of the predictors}
}
\value{
A function that takes a scalar or data.frame row and returns
a prediction.
}
\description{
Constructs a prediction function from a regression output
}
