% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{.gob}
\alias{.gob}
\title{Unsugared version of `gob`}
\usage{
.gob(Y, X, Z, pred.fn.ls, alpha = 0.95)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{A dataframe of covariates.}

\item{Z}{A binary vector of assignments.}

\item{pred.fn.ls}{A list of length 2.

The first element is the regression function to apply to the
control units, while the second element is the regression function
to apply to the treated units.}

\item{alpha}{The desired coverage of the confidence interval.}
}
\value{
a list of length 2. The first element is the point estimate;
the second is the confidence interval.
}
\description{
`.gob` does the actual work of computing the point estimate and
confidence interval. It should generally not be called directly
(and is in fact not exposed to the user).
}
