% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_to_string}
\alias{formula_to_string}
\title{Returns a string version of a formula}
\usage{
formula_to_string(form)
}
\arguments{
\item{form}{A formula}
}
\value{
a string representing the formula
}
\description{
Returns a string version of a formula
}
\examples{
\dontrun{
formula_to_string(Y ~ X + Y) # returns: "Y ~ X + Y"
}
}
