% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_predictor_names}
\alias{get_predictor_names}
\title{Extracts the names of the predictors in a formula}
\usage{
get_predictor_names(form)
}
\arguments{
\item{form}{A formula}
}
\value{
a vector of string representing the names of the predictors
}
\description{
Extracts the names of the predictors in a formula
}
\examples{
\dontrun{
get_predictor_names(Y ~ X + Z) # returns: c("X", "Z")
get_predictor_names(Y ~ X:C) # returns: c("X", "C")
get_predictor_names(Y ~ X + B + X:C) # returns: c("X", "B", "C")
}
}
