% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_dot_predictor_formula}
\alias{is_dot_predictor_formula}
\title{Checks whether the predictor side of a formula is a dot}
\usage{
is_dot_predictor_formula(form)
}
\arguments{
\item{form}{A formula}
}
\value{
TRUE if predictor side is a dot. FALSE otherwise.
}
\description{
Checks whether the predictor side of a formula is a dot
}
\examples{
\dontrun{
is_dot_predictor_formula(Y ~ .) # returns: TRUE
is_dot_predictor_formula(Y ~ X + Z) # returns: FALSE
}
}
