% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_one_predictor_formula}
\alias{is_one_predictor_formula}
\title{Checks whether the formula has a single predictor (which is not a dot)}
\usage{
is_one_predictor_formula(form)
}
\arguments{
\item{form}{A formula}
}
\value{
TRUE if formula has a single predictor, and that it is not a dot.
FALSE otherwise.
}
\description{
Checks whether the formula has a single predictor (which is not a dot)
}
\examples{
\dontrun{
is_one_predictor_formula(Y ~ X + Z) # returns: FALSE
is_one_predictor_formula(Y ~ .) # returns: FALSE
is_one_predictor_formula(Y ~ X) # returns: TRUE
}
}
