% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gob}
\alias{gob}
\title{Computes a point estimate and confidence interval using the GOB estimator}
\usage{
gob(form, Z, pred.fn.ls = NULL, data = NULL, alpha = 0.95)
}
\arguments{
\item{form}{A formula or a call-formula (see details for `call-formula`).}

\item{Z}{A vector of assignments.}

\item{pred.fn.ls}{A list of learners for the treatment and control potential
outcomes (see details).}

\item{data}{(optional) A dataframe whose columns include all the variables
referenced in `form`.

If `NULL`, the variables in `form` will be taken from the
calling environment.}

\item{alpha}{The confidence level (defaults to 0.95).}
}
\value{
A list whose first element is a point estimate, and second element
is a confidence interval.
}
\description{
Computes a point estimate and confidence interval using the GOB estimator
}
\examples{
library(RGOB)
N <- 100
X <- rnorm(N); B <- rnorm(N); C <- rnorm(N)
Z <- sample(c(0,1), N, replace=TRUE)
Y <- 1 + 0.2 * X + 0.5 * Z + rnorm(N, sd=0.1)
dat <- data.frame(Yp=Y, D=X, E=B, F=C)
gob(lm(Y ~ X + B), Z)
gob(lm(Yp~ D + E), Z, data=dat)
}
